pyproject.toml
[tool.poetry]
name = "API"
version = "0.1.0"
description = "City accessablity for inclusive people. BESTHackathon 2025 test assignment"
authors = ["Chugaister <dmytroshvahyliak@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
SQLAlchemy = "^2.0.32"
asyncpg = "^0.29.0"
pydantic = "^2.8.2"
starlette = "^0.37.2"
uvicorn = "^0.30.5"
python-dotenv = "^1.0.1"
alembic = "^1.13.2"
pyjwt = "^2.9.0"
pytz = "^2025.2"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
